add usage docs for context types and metadata, improve comments
