+++ metadata/headers.go
@@ -2,18 +2,18 @@
 package metadata
 
 var (
-	// HeaderTopic is the header name that contains topic name
+	// HeaderTopic is the header name that contains topic name.
 	HeaderTopic = "Micro-Topic"
-	// HeaderContentType specifies content type of message
+	// HeaderContentType specifies content type of message.
 	HeaderContentType = "Content-Type"
-	// HeaderEndpoint specifies endpoint in service
+	// HeaderEndpoint specifies endpoint in service.
 	HeaderEndpoint = "Micro-Endpoint"
-	// HeaderService specifies service
+	// HeaderService specifies service.
 	HeaderService = "Micro-Service"
-	// HeaderTimeout specifies timeout of operation
+	// HeaderTimeout specifies timeout of operation.
 	HeaderTimeout = "Micro-Timeout"
-	// HeaderAuthorization specifies Authorization header
+	// HeaderAuthorization specifies Authorization header.
 	HeaderAuthorization = "Authorization"
-	// HeaderXRequestID specifies request id
+	// HeaderXRequestID specifies request id.
 	HeaderXRequestID = "X-Request-Id"
 )
