+++ metadata/helpers.go
@@ -1,0 +1,7 @@
+package metadata
+
+func copyOf(v []string) []string {
+	vals := make([]string, len(v))
+	copy(vals, v)
+	return vals
+}
